+++ src/services/webSocket.sevice.ts
@@ -1,6 +1,7 @@
 import * as WebSocket from 'ws';
 import { Injectable, Logger } from '@nestjs/common';
 import { ESocketError } from 'src/constant/exception.constant';
+import { defer } from 'src/shared/utils';
 type Listener = (data: any) => void;
 @Injectable()
 export class WebSocketService {
